Add IncrementDecrementSvg component
